# Memory allocation optimization