@matklad 2020-05-16
bors r+

Thanks @simonvandel !

I believe there should be a lot of low hanging optimization fruit here.
If you are into these kinds of things, the most impactful thing to do would be to setup continuous benchmarking :)